+++ src/components/Select/select-list.tsx
@@ -1,5 +1,6 @@
 import classNames from "classnames";
-import type { CSSProperties, Ref } from "react";
+import { Check } from "lucide-react";
+import type { CSSProperties, FC, Ref } from "react";
 import { useJss } from "@/components/CopyUiProvider";
 import { useTheme } from "@/components/ThemeProvider";
 import styles from "./select-list.module.scss";
@@ -10,13 +11,22 @@ type SelectListProps<V extends string> = {
   style: CSSProperties;
 
   options: Array<{ value: V; label: string }>;
+  internalValue?: V | null;
   width: number;
 
   onItemClicked: (value: V) => void;
 };
 
 const SelectList = <V extends string>(props: SelectListProps<V>) => {
-  const { ref, className, style, options, width, onItemClicked } = props;
+  const {
+    ref,
+    className,
+    style,
+    options,
+    internalValue,
+    width,
+    onItemClicked,
+  } = props;
 
   const theme = useTheme();
   const jss = useJss();
@@ -48,22 +58,12 @@ const SelectList = <V extends string>(props: SelectListProps<V>) => {
     >
       {options.map((option) => {
         return (
-          <li key={option.value} className={styles.selectItem}>
-            <button
-              type="button"
-              className={styles.selectItemButton}
-              onClick={() => onItemClicked(option.value)}
-              onKeyDown={(e) => {
-                if (e.key === "Enter" || e.key === " ") {
-                  e.preventDefault();
-                  onItemClicked(option.value);
-                }
-              }}
-              tabIndex={0}
-            >
-              {option.label}
-            </button>
-          </li>
+          <SelectItem
+            key={option.value}
+            marked={option.value === internalValue}
+            label={option.label}
+            onChoose={() => onItemClicked(option.value)}
+          />
         );
       })}
     </ul>
@@ -71,5 +71,41 @@ const SelectList = <V extends string>(props: SelectListProps<V>) => {
 };
 
 SelectList.displayName = "Select.List";
+
+type SelectItemProps = {
+  marked: boolean;
+  label: string;
+  onChoose: () => void;
+};
+
+const SelectItem: FC<SelectItemProps> = (props) => {
+  const { marked, label, onChoose } = props;
+
+  const jss = useJss();
+  const theme = useTheme();
+  const stx = jss.hash({
+    "--select-item-mark-color": theme.colors.gray["600"],
+  });
+
+  return (
+    <li className={classNames(styles.selectItem, stx)}>
+      <button
+        type="button"
+        className={styles.selectItemButton}
+        onClick={() => onChoose()}
+        onKeyDown={(e) => {
+          if (e.key === "Enter" || e.key === " ") {
+            e.preventDefault();
+            onChoose();
+          }
+        }}
+        tabIndex={0}
+      >
+        <span>{label}</span>
+        {marked && <Check />}
+      </button>
+    </li>
+  );
+};
 
 export { SelectList };