+++ src/components/Select/CHANGELOG.md
@@ -11,3 +11,4 @@
 - 2025-10-28: Refactor code to make code cleaner. Add a small gap in trigger.
 - 2025-10-30: Add a `ResizeObserver` to the `Select.Trigger` to update the
   `Select.List`'s width when the trigger size changes.
+- 2025-10-31: Add the check mark for the selected item in the list.